change feature: approved users were changed of the file
