+++ src/app/admin-dashboard/page.js
@@ -9,10 +9,7 @@ import styles from '../../styles/admin-dashboard/page.module.css'
 
 export default function AdminDashboard() {
   const [users, setUsers] = useState([]);
-  const [approvedUsers, setApprovedUsers] = useState([]);
   const [message, setMessage] = useState('');
-  const [selectedUser, setSelectedUser] = useState(null);
-  const [workoutForm, setWorkoutForm] = useState({ description: '', date: '' });
   const router = useRouter();
 
   useEffect(() => {
@@ -33,52 +30,13 @@ export default function AdminDashboard() {
       } catch (error) {
         console.error('Erro ao buscar usuários pendentes:', error);
         setMessage('Erro ao processar a requisição.');
-      }
-    };
-
-    const fetchApprovedUsers = async () => {
-      try {
-        const response = await fetch('/api/list-users');
-        if (response.ok) {
-          const data = await response.json();
-          setApprovedUsers(data);
-        } else {
-          console.error('Erro ao buscar usuários aprovados');
-        }
-      } catch (error) {
-        console.error('Erro ao buscar usuários aprovados:', error);
-      }
+      };
     };
 
     fetchUsers();
-    fetchApprovedUsers();
   }, [router]);
 
-  const handleApprove = async (email) => {
-    try {
-      const response = await fetch('/api/approve-user', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ email }),
-      });
 
-      const data = await response.json();
-
-      if (response.ok) {
-        setMessage('Usuário aprovado com sucesso.');
-        setUsers(users.filter(user => user.email !== email));
-
-        setApprovedUsers([...approvedUsers, users.find(user => user.email === email)]);
-      } else {
-        setMessage(data.message);
-      }
-    } catch (error) {
-      console.error('Erro ao aprovar usuário:', error);
-      setMessage('Erro ao aprovar o usuário.');
-    }
-  };
 
   const handleLogout = async () => {
     try {
@@ -90,46 +48,6 @@ export default function AdminDashboard() {
     } catch (error) {
       console.error('Erro ao fazer logout:', error);
       setMessage('Erro ao realizar logout.');
-    }
-  };
-
-  const handleSelectUser = (user) => {
-    setSelectedUser(user);
-  };
-
-  const handleWorkoutFormChange = (e) => {
-    setWorkoutForm({ ...workoutForm, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmitWorkout = async (e) => {
-    e.preventDefault();
-    if (!selectedUser) {
-      setMessage('Selecione um usuário primeiro.');
-      return;
-    }
-
-    try {
-      const response = await fetch('/api/create-workout', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          userId: selectedUser._id,
-          workoutData: workoutForm
-        }),
-      });
-
-      if (response.ok) {
-        setMessage('Treino criado com sucesso.');
-        setWorkoutForm({ description: '', date: '' });
-      } else {
-        const errorData = await response.json();
-        setMessage(errorData.message || 'Erro ao criar treino.');
-      }
-    } catch (error) {
-      console.error('Erro ao criar treino:', error);
-      setMessage('Erro ao processar a requisição.');
     }
   };
 
@@ -156,50 +74,6 @@ export default function AdminDashboard() {
             <p>Não há usuários pendentes.</p>
           )}
         </ul>
-
-        <h2>Usuários Aprovados</h2>
-        <ul>
-          {approvedUsers.length > 0 ? (
-            approvedUsers.map(user => (
-              <li key={user.email}>
-                {user.nome} - {user.email}
-                <button onClick={() => handleSelectUser(user)}>Selecionar para Treino</button>
-              </li>
-            ))
-          ) : (
-            <p>Não há usuários aprovados.</p>
-          )}
-        </ul>
-
-        {selectedUser && (
-          <div>
-            <h2>Criar Treino para {selectedUser.nome}</h2>
-            <form onSubmit={handleSubmitWorkout}>
-              <div>
-                <label htmlFor="description">Descrição do Treino:</label>
-                <textarea
-                  id="description"
-                  name="description"
-                  value={workoutForm.description}
-                  onChange={handleWorkoutFormChange}
-                  required
-                />
-              </div>
-              <div>
-                <label htmlFor="date">Data do Treino:</label>
-                <input
-                  type="date"
-                  id="date"
-                  name="date"
-                  value={workoutForm.date}
-                  onChange={handleWorkoutFormChange}
-                  required
-                />
-              </div>
-              <button type="submit">Criar Treino</button>
-            </form>
-          </div>
-        )}
       </div>
     </div>
   );
